requirements.txt
pillow==1.7.8
git+http://github.com/pybrain/pybrain.git#egg=pybrain
